refactor: move AnalysisRunner out of validation package
