+++ src/main/java/org/variantsync/diffdetective/validation/EditClassValidation.java
@@ -1,5 +1,6 @@
 package org.variantsync.diffdetective.validation;
 
+import org.variantsync.diffdetective.AnalysisRunner;
 import org.variantsync.diffdetective.analysis.Analysis;
 import org.variantsync.diffdetective.analysis.FilterAnalysis;
 import org.variantsync.diffdetective.analysis.PreprocessingAnalysis;
